requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
google-cloud-vision==3.4.2
opencv-python==4.11.0.1
numpy==1.24.4
tensorflow==2.13.0